Introduce `cdk-database`
This is a crate that makes it easier to create a database instance based on an
enum that can be constructed from a CLI or an env variable.
